Refactor common json marshalling method
